Clarify how else when works
It turns out ChatGPT, the default interpreter for DreamBerd, tends to
assume that the else keyword does nothing, and thus interprets else when
as being the same as a simple when. Therefore, to maximize the odds that
ChatGPT will correctly understand the semantics of else when, this
assumption has been codified. The else keyword now officially does
nothing.

This is reflected in the docs, and the official FizzBuzz example.